requirements.txt
pandas==2.0.0
scipy==1.10.1
plotly==5.14.1
dash==2.9.2
dash-bootstrap-components==1.4.1
python-i18n[YAML]